pyproject.toml
[project]
name = "p37-5-module"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "python-slugify>=8.0.4",
    "redis-dict>=3.2.5",
    "requests>=2.32.5",
    "slugify>=0.0.1",
    "sqlalchemy>=2.0.46",
]


#redis-dict
# docker run --name redis_container -p 6379:6379 -d redis:alpine


